marginally better of stdout + stderr display
bazel sends stdout and stderr separately, when in reality they are intertwined. if you put the stdout after the stderr, the next stderr batch can ssend control codes meant to wipe the stderr, but actually wipes the stderr. The downside of putting stdout first is that output can appear slightly out of order, but at least it won't get wiped.